+++ test/graphio_test.cpp
@@ -12,7 +12,7 @@ TEST(GraphIOTest, exportGML) {
 TEST(GraphIOTest, importGML) {
     UndirectedGraph g = GraphImporter::importGML<UndirectedGraph>("graphio1.gml");
     EXPECT_EQ(g.getOrder(), 9);
-    for (uint u = 0; u < g.getOrder()-1; u++) {
+    for (uint64_t u = 0; u < g.getOrder()-1; u++) {
         EXPECT_EQ(g.deg(u)/2, 2);
     }
     EXPECT_EQ(g.deg(g.getOrder()-1)/2, 8);
